Separate (de)serialization from metadata class. Cleanups.
